Finished the translation for a file: "hangman.py"->"ahorcado.py" Added my name and contact info in "SPECIAL_THANKS_TO.txt".
NOTE 1: I took the liberty to remove the accents from the animal names. The alternatives would be either: a) adding the characters á, é, í, ó, ú to the dictionary and considering a!=á (not intuitive for a Spanish speaker) , or b) modifying the code, which goes against the instructions in the readme file.

NOTE 2: I could not find (either at http://docs.python.org.ar/tutorial/ or somewhere else) an accepted Spanish translation for the words "set" and "get" in method names. These words are most often kept in English. For the rest of this book I intend to stick to the following: set->establecer; get->obtener